Refactor dependency tracing behing an interface
